Transfer constants in a separate file AppConstants
