+++ src/constants/AppConstants.js
@@ -1,0 +1,5 @@
+export const SET_LINKS = 'SET_LINKS';
+export const ADD_LINK = 'ADD_LINK';
+export const LINK_FETCHED = 'LINK_FETCHED';
+export const LINK_UPDATED = 'LINK_UPDATED';
+export const LINK_DELETED = 'LINK_DELETED';
