+++ src/reducers/links.js
@@ -1,7 +1,10 @@
-import { SET_LINKS, LINK_FETCHED, LINK_UPDATED } from '../actions/AppActions';
+import { SET_LINKS, LINK_FETCHED, LINK_UPDATED, LINK_DELETED } from '../constants/AppConstants';
 
 export default function links(state = [], action = {}) {
   switch (action.type) {
+    case LINK_DELETED:
+      return state.filter(item => item._id !== action.linkId);
+    
     case LINK_UPDATED:
       return state.map(item => {
         if (item._id === action.link._id) return action.link;
